propogate events triggered in item.editor.render
Any events triggered in the editor's render function are not propogated to this list editor. For example, the 'open' trigger for editors that use Modals is never caught. This commit separtes the creation of the item / editor and the rendering of the item / editor. After creating the item and editor, events can be bound to item.editor before proceeding with the rendering (which will trigger events).